Move replication storage preparation to node
